+++ apps/pwabuilder/Models/ServiceWorkerResult.cs
@@ -23,4 +23,15 @@ namespace PWABuilder.Models
         public int Status { get; set; }
         public ServiceWorkerBodyResult Body { get; set; }
     }
+
+    public class AnalyzeServiceWorkerResponse
+    {
+        public bool? DetectedBackgroundSync { get; set; }
+        public bool? DetectedPeriodicBackgroundSync { get; set; }
+        public bool? DetectedPushRegistration { get; set; }
+        public bool? DetectedSignsOfLogic { get; set; }
+        public bool? DetectedEmpty { get; set; }
+        public List<string>? Raw { get; set; }
+        public string? Error { get; set; }
+    }
 }
